+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Minor changes and fixes
 
-* Anonymous chat user: remember the chosen nickname in sessionStorage, to avoid entering it again too often.
-* Fix: if an anonymous chat user enter spaces in the nickname choice, it will allows them to keep the random nickname.
+* Anonymous chat user: remember the chosen nickname in localStorage, to avoid entering it again too often. Nickname will expire after 12 hours.
+* Fix: if an anonymous chat user enter spaces in the nickname choice, it allowed them to keep the random nickname.
 * Authenticated users: if current user nickname is already used in the room, automatically add a suffix.
 * UX: add a label ('Choose a nickname to enter') for the anonymous nickname prompt. Fix #287.
 * Translation updates: German, French.